backend/requirements.txt
APScheduler==3.7.0
asgiref==3.3.1
astroid==2.5
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.7
django-apscheduler==0.5.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
flake8==3.8.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
Markdown==3.3.3
mccabe==0.6.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.20.1
pandas==1.2.2
protobuf==3.15.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.3
wrapt==1.12.1
